Task 6/requirements.txt
streamlit
requests
Pillow
textblob
python-dotenv
